Fix Android build.

+++ filesystem/filesystem.cpp
@@ -319,7 +319,7 @@ struct ScratchFilesystemFile final : Internal::File
 	{
 	}
 
-	size_t get_size() override
+	uint64_t get_size() override
 	{
 		return data.size();
 	}
